% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFlowGraph.R
\name{createFlowGraph}
\alias{createFlowGraph}
\title{igraph object from flow direction raster}
\usage{
createFlowGraph(fdirpath, discharge = NULL)
}
\arguments{
\item{discharge}{optional: raster object with discharge values}

\item{fdir}{a raster object with flow directions (1,2,4,8,16,32,64,128)}
}
\value{
object of class igraph with flow network topology
}
\description{
Creates an igraph graph from a flow direction raster
}
\examples{
\dontrun{
library(raster)
data("volcano")
v <- raster(volcano, xmn=2667400, xmx=2668010, ymn=6478700, ymx=6479570, crs="+init=epsg:27200")
fd <- terrain(v, opt = "flowdir")
fdGraph <- GRDCFlowTools::createFlowGraph(fd)
}
}
