% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSplitFlowGraph.R
\name{createSplitFlowGraph}
\alias{createSplitFlowGraph}
\title{igraph object from flow direction raster (splitted)}
\usage{
createSplitFlowGraph(fdirpath, ny = 10, rootpath = NULL)
}
\arguments{
\item{fdirpath}{path to a raster object with flow directions (1,2,4,8,16,32,64,128)}

\item{ny}{number of strips the raster will be splitted into}

\item{rootpath}{path to directory where igraph objects will be written to}
}
\value{
list of named list (slot path: path to graph; slot nmin: minimum cell number of this graph; slot nmax: maximum cell number of this graph)
}
\description{
Creates a list of igraph objects from a splitted flow direction raster
}
\examples{
\dontrun{
library(raster)
data("volcano")
v <- raster(volcano, xmn=2667400, xmx=2668010, ymn=6478700, ymx=6479570, crs="+init=epsg:27200")
fd <- terrain(v, opt = "flowdir")
tempfilename <- paste0(tempfile(),".asc")
raster::writeRaster(fd, filename = tempfilename)
rl <- GRDCFlowTools::createSplitFlowGraph(fdirpath = tempfilename, ny = 20, rootpath = tempdir())
}
}
