% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasinCells.R
\name{getBasinCells}
\alias{getBasinCells}
\title{get Basin Cells from Flow Graph}
\usage{
getBasinCells(cell, flowgraph)
}
\arguments{
\item{cell}{cell number}

\item{flowgraph}{igraph object with flow network topology}
}
\value{
vector of cell numbers
}
\description{
Derive all inflow cells to a given cell number of a flow direction raster
}
\examples{
\dontrun{
library(raster)
data("volcano")
v <- raster(volcano, xmn=2667400, xmx=2668010, ymn=6478700, ymx=6479570, crs="+init=epsg:27200")
fd <- terrain(v, opt = "flowdir")
fdGraph <- GRDCFlowTools::createFlowGraph(fd)
getBasin <- GRDCFlowTools::getBasinCells(4123, fdGraph)
}
}
