% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasinCellsSplitGraph.R
\name{getBasinFromSplitGraph}
\alias{getBasinFromSplitGraph}
\title{get Basin Cells from Split Flow Graph}
\usage{
getBasinFromSplitGraph(cell, resultList, index = NULL)
}
\arguments{
\item{cell}{cell number}

\item{resultList}{list returned by function createSplitFlowGraph}

\item{index}{force function to search in graph with this index}
}
\value{
vector of cell numbers
}
\description{
Derive all inflow cells to a given cell number of a splitted flow direction raster
}
\examples{
\dontrun{
library(raster)
data("volcano")
fd <- terrain(v, opt = "flowdir")
tempfilename <- paste0(tempfile(),".asc")
writeRaster(fd, tempfilename)
fdSplitGraph <- GRDCFlowTools::createSplitFlowGraph(tempfilename, ny = 2, rootpath = tempdir())
getBasin <- GRDCFlowTools::getBasinFromSplitGraph(4123, fdSplitGraph)
}
}
