% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWatershed.R
\name{getWatershed}
\alias{getWatershed}
\title{get Basin from Flow Graph}
\usage{
getWatershed(cell, fdir)
}
\arguments{
\item{cell}{cell number}

\item{fdir}{a raster object with flow directions (1,2,4,8,16,32,64,128)}
}
\value{
shapefile as object class sf
}
\description{
derive and create a shapefile of all inflow cells to a given cell number of a flow direction raster
}
\examples{
\dontrun{
library(raster)
data("volcano")
v <- raster(volcano, xmn = 2667400, xmx = 2668010, ymn = 6478700, ymx = 6479570, crs="+init=epsg:27200")
fd <- terrain(v, opt = "flowdir")
fdGraph <- GRDCFlowTools::createFlowGraph(fd)
getCells <- GRDCFlowTools::getBasinCells(4123,fdGraph)
watershed <- GRDCFlowTools::getWatershed(getCells, fd)
}
}
